Move algorithm_test_utils header in execution_base
