+++ libs/pika/async_cuda/tests/unit/then_with_stream.cu
@@ -11,7 +11,7 @@
 
 #include <whip.hpp>
 
-#include "algorithm_test_utils.hpp"
+#include <pika/execution_base/tests/algorithm_test_utils.hpp>
 
 #include <atomic>
 #include <cstddef>
