+++ libs/pika/execution/tests/CMakeLists.txt
@@ -8,9 +8,6 @@ include(pika_message)
 include(pika_option)
 
 if(PIKA_WITH_TESTS)
-  add_library(pika_execution_test_utilities INTERFACE)
-  target_include_directories(pika_execution_test_utilities INTERFACE include/)
-
   if(PIKA_WITH_TESTS_UNIT)
     pika_add_pseudo_target(tests.unit.modules.execution)
     pika_add_pseudo_dependencies(
