+++ libs/pika/execution/tests/unit/algorithm_let_error.cpp
@@ -7,7 +7,7 @@
 #include <pika/modules/execution.hpp>
 #include <pika/testing.hpp>
 
-#include "algorithm_test_utils.hpp"
+#include <pika/execution_base/tests/algorithm_test_utils.hpp>
 
 #include <atomic>
 #include <exception>
