+++ libs/pika/execution/tests/unit/algorithm_when_all_vector.cpp
@@ -8,7 +8,7 @@
 #include <pika/modules/execution.hpp>
 #include <pika/testing.hpp>
 
-#include "algorithm_test_utils.hpp"
+#include <pika/execution_base/tests/algorithm_test_utils.hpp>
 
 #include <atomic>
 #include <cstddef>
